Bifurcate the expected error message based on runtime mode (#2821)
## Motivation and Context
Makes a failing test `multi_region_access_points` pass in the
orchestrator

## Description
The orchestrator has already an `AwsAuthStage` counterpart implemented
(which is where the expected error for `multi_region_access_points`
comes from in the middleware mode).

This PR just updates the expected error message in the test for the
orchestrator to make it pass.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._

Co-authored-by: Yuki Saito <awsaito@amazon.com>