Update Kotlin unit tests
This commit updates `config_override`-related unit tests written in
Kotlin so that the new way of testing should check the behavior of each
trait object retrieved from a config override layer. Previously, the
tests were written in a naive way in that it only checks whether the
appropriate orchestrator component was inserted into a config override
layer when a test calls a certain builder method.