Move endpoint module and add endpoint re-exports (#2822)
All the types in the `crate::endpoint` module are related to endpoint
configuration, so they should be moved into `crate::config::endpoint` to
be consistent with the rest of the generated crate organization. This PR
moves them for the orchestrator implementation, but retains them in
`crate::endpoint` when generating for middleware.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._