Fix flag name to retrieve if runtime plugins are configurable (#2885)
## Motivation and Context
Follow-up on #2864. It got it 99% right, it's just that the name of a
flag for retrieving its value was incorrect.

## Testing
Generated an AWS SDK and confirmed that setters on `Config` for runtime
plugins are marked as `pub(crate)` correctly.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._

Co-authored-by: ysaito1001 <awsaito@amazon.com>