Add support for error-correcting builders (#2991)
## Motivation and Context
To implement #1767 we need to support error-correction to default values
when instantiating builders.

-
https://smithy.io/2.0/spec/aggregate-types.html?highlight=error%20correction#client-error-correction
## Description
Adds `pub(crate) correct_errors_<shape>` method that will be used in
deserialization to set default values for required fields when not set
in the serialized response. This only applies to client via
`ClientBuilderInstantiator`

## Testing
- added a new test that's fairly exhaustive

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [ ] I have updated `CHANGELOG.next.toml` if I made changes to the
smithy-rs codegen or runtime crates
- [ ] I have updated `CHANGELOG.next.toml` if I made changes to the AWS
SDK, generated SDK code, or SDK runtime crates

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._