Update a module for the last release date in canary (#3013)
## Motivation and Context
Fixes [canary
failures](https://github.com/awslabs/aws-sdk-rust/actions/runs/6345490441/job/17237658244)
by updating the-last-release-before-breaking-change module in
`canary-lambda`.

## Description
[A previous PR](https://github.com/awslabs/smithy-rs/pull/2978) removing
the `Steram` trait from public API had a breaking change, requiring
canary to update its code. The latest release at the time while the PR
was worked on was `release-2023-08-23` so we created a new module with
that date and moved the previously working canary's code prior to the
breaking change to it. The `latest` module then contained canary's code
after the breaking change.

However, before that PR was merged to main, we made several releases in
`aws-sdk-rust`, making `release-2023-08-23` no longer the latest release
before the breaking change; a new one is `release_2023_09_25`. Hence
this PR.

## Testing
Manually executed
```
➜ canary-runner git:(ysaito/fix-canary-module-dates) cargo run -- build-bundle --canary-path ../canary-lambda --sdk-release-tag release-2023-09-25 --musl --manifest-only
➜ canary-runner git:(ysaito/fix-canary-module-dates) cd ../canary-lambda
➜ canary-lambda git:(ysaito/fix-canary-module-dates) cargo check
```
(also tested for `--sdk-release-tag` with `release-2023-09-22` and
`release-2023-09-20`)

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._