Make versions comparison more flexible in the fix-manifest subcommand of publisher (#3025)
## Motivation and Context
Attempts to resolve [a workflow
failure](https://github.com/awslabs/aws-sdk-rust/actions/runs/6395300944/job/17358637767#step:6)
in `aws-sdk-rust`.

## Description
We have observed a series of `WARN`s from the `check-manifests` job. It
is caused by #3009, specifically `update_dep` in the `fix-manifests`
needs to be updated where it compares a version in a `Table` versus a
current version being looked at. For instance, prior to the PR, `0.56.1`
and `~0.56` for a crate were considered a mismatch when it should be a
match according to our latest logic.

## Testing
- Added a unit test for a helper function.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._