Enable the invoke_model_with_response_stream API (#3031)
This API is a streaming API—I tested it manually, an E2E test an example
are forthcoming. I also removed the unused `config` service from the
smoketest and added bedrock-runtime.

## Motivation and Context
This API is suppressed by default because it's an event stream API

## Description


## Testing
https://gist.github.com/rcoh/ce61429303a56fefd65ba05bd36d28fc works



----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._