fix Publisher Workflow to successfully create new branches
The publisher tool was unable to create new branches because it attempted to determine if the current commit was on main. However, this didn't work! Since main was never checked out (ever) in the repo, there was no ref to look for. This fixes the issue in (perhaps not the best) way by simply checking out main first.