fix Publisher Workflow to successfully create new branches (#3117)
The publisher tool was unable to create new branches because it
attempted to determine if the current commit was on main. However, this
didn't work! Since main was never checked out (ever) in the repo, there
was no ref to look for. This fixes the issue in (perhaps not the best)
way by simply checking out main first.

The logic was also refactored to have dry run and non-dry run behave in
a way that is closer to identical.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._