Update Canary OIDC provider, NPM commands, and previous version pagination test (#3142)
The OIDC provider no longer needs to specify the thumbprints now; In the
IAM Identity Provider console it says that those are now sourced
automatically.
I also updated the npm commands to build and deploy everything instead
of just the canary and I added a trait import to the previous version
canary in the hope of getting past that warning that only appears in CI.
If that doesn't fix it, then Yuki will when he takes a look next week.

- update canary npm commands
- update canary OIDC provider stack
- add missing trait to previous version canary

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._